Fixes for compilation after renaming.
