+++ k2/python/csrc/torch/arc.cu
@@ -24,8 +24,8 @@
 
 #include "k2/csrc/device_guard.h"
 #include "k2/csrc/fsa.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/arc.h"
-#include "k2/python/csrc/torch/torch_util.h"
 
 namespace k2 {
