+++ k2/python/csrc/torch/index_select.cu
@@ -32,8 +32,8 @@
 #include "k2/csrc/ragged.h"
 #include "k2/csrc/ragged_ops.h"
 #include "k2/csrc/tensor_ops.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/index_select.h"
-#include "k2/python/csrc/torch/torch_util.h"
 #include "k2/python/csrc/torch/v2/ragged_any.h"
 
 namespace k2 {
