+++ k2/python/csrc/torch/mutual_information.cu
@@ -19,8 +19,8 @@
  */
 
 #include "k2/csrc/device_guard.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/mutual_information.h"
-#include "k2/python/csrc/torch/torch_util.h"
 
 void PybindMutualInformation(py::module &m) {
   m.def(
