+++ k2/python/csrc/torch/rnnt_decode.cu
@@ -29,8 +29,8 @@
 #include "k2/csrc/device_guard.h"
 #include "k2/csrc/fsa.h"
 #include "k2/csrc/rnnt_decode.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/rnnt_decode.h"
-#include "k2/python/csrc/torch/torch_util.h"
 
 namespace k2 {
 static void PybindRnntDecodingConfig(py::module &m) {
