+++ k2/python/csrc/torch/v2/autograd/index_and_sum.h
@@ -25,7 +25,7 @@
 
 #include "k2/csrc/ragged_ops.h"
 #include "k2/csrc/tensor_ops.h"
-#include "k2/python/csrc/torch/torch_util.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/v2/ragged_any.h"
 
 namespace k2 {
