+++ k2/python/csrc/torch/v2/autograd/sum.h
@@ -26,7 +26,7 @@
 #include <utility>
 
 #include "k2/csrc/ragged_ops.h"
-#include "k2/python/csrc/torch/torch_util.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/v2/ragged_any.h"
 
 namespace k2 {
