+++ k2/python/csrc/torch/v2/ragged_shape.cu
@@ -25,7 +25,7 @@
 #include "k2/csrc/device_guard.h"
 #include "k2/csrc/ragged.h"
 #include "k2/csrc/ragged_ops.h"
-#include "k2/python/csrc/torch/torch_util.h"
+#include "k2/csrc/torch_util.h"
 #include "k2/python/csrc/torch/v2/doc/ragged_shape.h"
 #include "k2/python/csrc/torch/v2/ragged_shape.h"
